inicio comunicacion del planificador

+++ planificador/src/algoritmos.c
@@ -3,7 +3,7 @@ double calcularProximaRafaga(double estimadoAnterior, double realAnterior)
 	return estimadoAnterior*0.5 + realAnterior*0.5;
 }
 
-float calcularRR(double tEnEspera, double estimadoAnterior, double realAnterior)
+double calcularRR(double tEnEspera, double estimadoAnterior, double realAnterior)
 {
 	return (1+ tEnEspera/calcularProximaRafaga(estimadoAnterior,realAnterior));
 }
